+++ src/logging/index.ts
@@ -1,7 +1,5 @@
+import {ApiService} from '../api/api-service';
 import {PlayerLoggerService} from './player-logger-service';
 import {PlayerDeviceTypes} from '../models/player';
 
-export {
-    PlayerLoggerService,
-    PlayerDeviceTypes,
-};
+export {ApiService, PlayerLoggerService, PlayerDeviceTypes};