+++ webpack.config.js
@@ -1,14 +1,13 @@
 import path from 'path';
 import {fileURLToPath} from 'url';
 import WebpackConcatPlugin from 'webpack-concat-files-plugin';
-import CopyPlugin from 'copy-webpack-plugin';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 export default {
     entry: {
         bundle: './src/index.ts',
-        logging: './src/logging/index.ts'
+        logging: './src/logging/index.ts',
     },
     module: {
         rules: [